fix module inception for dime and storage; remove temp lint override
